requirements.txt
numpy
openai
langchain
python-dotenv